[CHI-89] Add dropdown support to Tabs Code refactoring just before removing vertical support for Navigation component.
