+++ src/chi/javascript/components/auxiliary/event-manager.js
