+++ src/chi/javascript/components/auxiliary/sliding-border.js
@@ -1,4 +1,4 @@
-import {Util} from "../core/util.js";
+import {Util} from "../../core/util.js";
 
 const CLASS_SLIDING_BORDER='a-sliding-border';
 const CLASS_SLIDING='-sliding';
